web_scrape/requirements.txt
pydantic==2.10.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dateutil==2.9.0
python-dotenv==1.0.1
spacy==3.7.5
fi_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fi_core_news_sm-3.8.0/fi_core_news_sm-3.8.0-py3-none-any.whl
beautifulsoup4==4.12.3
typing-extensions==4.12.2
langdetect==1.0.9
deep-translator==1.11.4
aiohttp==3.11.11
